# docs/openapi.yaml
openapi: 3.0.3
info:
  title: RAMP UP — Engineering Onboarding Portal API
  description: |
    Tài liệu đặc tả API (OpenAPI Specification) cho hệ thống **RAMP UP** (Glinteco e-Learning).
    Hệ thống cung cấp cổng thông tin đào tạo nội bộ gamified dành cho các kỹ sư mới gia nhập Glinteco.
    
    ### Quy ước chung:
    * **Base URL:** `/api/v1` (mặc định môi trường development: `https://api.glinteco-elearning.dev/api/v1`)
    * **Xác thực:** Ngoại trừ các route Auth public, tất cả các request yêu cầu header `Authorization: Bearer <accessToken>`.
    * **Định dạng thời gian:** Chuẩn ISO-8601 UTC (ví dụ: `2026-06-16T15:16:56Z`).
    * **Phân trang danh sách (Pagination):** Sử dụng Keyset/Cursor Pagination với định dạng phản hồi chung:
      ```json
      {
        "data": [...],
        "nextCursor": "<opaque_string_or_null>",
        "hasMore": true
      }
      ```
    * **Quản lý lỗi (Error Envelopes):** Các lỗi trả về định dạng chuẩn:
      ```json
      {
        "statusCode": 400,
        "message": ["email must be a valid email"],
        "error": "Bad Request"
      }
      ```
  version: 1.0.0
servers:
  - url: https://api.glinteco-elearning.dev/api/v1
    description: Môi trường Development
paths:
...
  /exercises/{id}/submissions:
    post:
      summary: Nộp link bài làm PR lần đầu (Learner)
      description: Học viên nộp bài tập bằng cách gửi link PR GitHub. Bài nộp được khởi tạo ở trạng thái `submitted` để chờ Admin review.
      tags:
        - Submissions
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          description: ID của bài tập
          required: true
          schema:
            type: string
            example: e3
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - prUrl
              properties:
                prUrl:
                  type: string
                  description: URL của Pull Request GitHub chứa bài giải
                  example: github.com/acme/web/pull/468
      responses:
        '201':
          description: Nộp bài thành công.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Submission'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          description: Không tìm thấy bài tập.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      summary: Nộp lại / Cập nhật bài giải PR (Learner)
      description: Học viên cập nhật lại link PR khác hoặc cập nhật lại PR sau khi có yêu cầu sửa đổi từ Admin. Trạng thái sẽ đặt lại thành `submitted`.
      tags:
        - Submissions
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          description: ID bài tập
          required: true
          schema:
            type: string
            example: e3
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - prUrl
              properties:
                prUrl:
                  type: string
                  example: github.com/acme/web/pull/470
      responses:
        '200':
          description: Cập nhật bài nộp thành công.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Submission'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /submissions:
    get:
      summary: Hàng đợi chấm bài / Lịch sử bài nộp (Admin)
      description: Lấy danh sách bài nộp phục vụ cho giao diện review chấm bài của Admin. Mặc định dùng để lấy danh sách bài đang chờ duyệt (`?status=submitted`).
      tags:
        - Submissions
      security:
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          description: Lọc theo trạng thái bài nộp
          required: false
          schema:
            type: string
            enum: [submitted, approved, changes]
            example: submitted
        - name: cohortId
          in: query
          description: Lọc bài nộp của học viên trong cohort cụ thể
          required: false
          schema:
            type: string
            example: c_spring2026
        - name: limit
          in: query
          description: Giới hạn bản ghi trả về
          required: false
          schema:
            type: integer
            maximum: 50
            default: 20
        - name: cursor
          in: query
          description: Cursor phân trang danh sách
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Thành công.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/SubmissionFeedItem'
                  nextCursor:
                    type: string
                    nullable: true
                    example: eyJzdWJtaXR0ZWRBdCI6…
                  hasMore:
                    type: boolean
                    example: true
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /submissions/mine:
    get:
      summary: Danh sách bài tập đã nộp của tôi (Learner)
      description: Trả về danh sách tất cả các bài nộp của chính learner đang đăng nhập.
      tags:
        - Submissions
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Thành công.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          example: s1
                        exerciseId:
                          type: string
                          example: e2
                        exercise:
                          type: string
                          example: Add a Paginated Users Endpoint
                        prUrl:
                          type: string
                          example: github.com/acme/api/pull/119
                        status:
                          type: string
                          enum: [submitted, approved, changes]
                          example: submitted
                        submittedAt:
                          type: string
                          format: date-time
                          example: 2026-06-15T11:11:00Z
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /submissions/{id}:
    get:
      summary: Xem chi tiết một bài nộp (Authenticated)
      description: |
        Xem chi tiết một bản ghi bài nộp cùng trạng thái và nhận xét.
        * Phân quyền:* Chủ nhân bài nộp (`learner`) hoặc bất kỳ `admin` nào được quyền truy cập.
      tags:
        - Submissions
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          description: ID của bài nộp (Submission ID)
          required: true
          schema:
            type: string
            example: s1
      responses:
        '200':
          description: Thành công.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubmissionDetail'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          description: Không tìm thấy bài nộp.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /submissions/{id}/approve:
    post:
      summary: Phê duyệt bài nộp (Admin)
      description: Admin phê duyệt bài giải PR của học viên. Bài tập sẽ chuyển sang trạng thái `approved`, học viên nhận được XP thưởng của bài tập và tổng XP của họ được cập nhật.
      tags:
        - Submissions
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          description: ID của bài nộp
          required: true
          schema:
            type: string
            example: s1
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                note:
                  type: string
                  description: Nhận xét phê duyệt từ Admin
                  example: Clean keyset impl, nice tests.
      responses:
        '200':
          description: Phê duyệt thành công.
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    example: s1
                  status:
                    type: string
                    example: approved
                  exerciseId:
                    type: string
                    example: e2
                  exerciseStatus:
                    type: string
                    example: approved
                  reviewerId:
                    type: string
                    example: u_lead
                  reviewedAt:
                    type: string
                    format: date-time
                    example: 2026-06-15T11:50:00Z
                  xpAwarded:
                    type: integer
                    example: 200
                  learnerTotalXp:
                    type: integer
                    example: 1440
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          description: Không tìm thấy.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /submissions/{id}/request-changes:
    post:
      summary: Yêu cầu sửa đổi bài nộp (Admin)
      description: Admin từ chối phê duyệt và yêu cầu học viên sửa lại bài làm kèm theo ghi chú nhận xét cụ thể. Bài nộp sẽ chuyển sang trạng thái `changes`.
      tags:
        - Submissions
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          description: ID bài nộp
          required: true
          schema:
            type: string
            example: s2
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - note
              properties:
                note:
                  type: string
                  description: Nội dung chi tiết các yêu cầu học viên cần chỉnh sửa
                  example: Status-dot color should reuse getStatusColor() instead of a new switch.
      responses:
        '200':
          description: Yêu cầu sửa đổi thành công.
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    example: s2
                  status:
                    type: string
                    example: changes
                  exerciseId:
                    type: string
                    example: e1
                  exerciseStatus:
                    type: string
                    example: pending
                  reviewerId:
                    type: string
                    example: u_lead
                  reviewNote:
                    type: string
                    example: Status-dot color should reuse getStatusColor()…
                  reviewedAt:
                    type: string
                    format: date-time
                    example: 2026-06-15T11:52:00Z
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /submissions/{id}/history:
    get:
      summary: Tra cứu lịch sử các lần nộp của bài tập (Proposed - Authenticated)
      description: Lấy nhật ký tất cả các phiên bản nộp bài (các link PR cũ) và nội dung nhận xét tương ứng qua các vòng review chấm bài của bài tập này.
      tags:
        - Submissions
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          description: ID của bài nộp chính
          required: true
          schema:
            type: string
            example: s1
      responses:
        '200':
          description: Thành công.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubmissionHistoryResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          description: Không tìm thấy lịch sử.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # =========================================================================
  # 8. SEARCH & NOTIFICATIONS
  # =========================================================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Gửi access token dạng 'Bearer <Token>' trong header Authorization.
  responses:
    ValidationError:
      description: Dữ liệu đầu vào không hợp lệ (lỗi validate DTO).
      content:
        application/json:
          schema:
            type: object
            properties:
              statusCode:
                type: integer
                example: 400
              message:
                type: array
                items:
                  type: string
                example: ["email must be a valid email", "password must be longer than or equal to 6 characters"]
              error:
                type: string
                example: Bad Request
    UnauthorizedError:
      description: Chưa đăng nhập hoặc Token hết hạn/không hợp lệ.
      content:
        application/json:
          schema:
            type: object
            properties:
              statusCode:
                type: integer
                example: 401
              message:
                type: string
                example: Unauthorized
              error:
                type: string
                example: Unauthorized
    ForbiddenError:
      description: Người dùng không có quyền truy cập chức năng này (sai role).
      content:
        application/json:
          schema:
            type: object
            properties:
              statusCode:
                type: integer
                example: 403
              message:
                type: string
                example: Forbidden resource
              error:
                type: string
                example: Forbidden
  schemas:
    ErrorResponse:
      type: object
      properties:
        statusCode:
          type: integer
          example: 400
        message:
          type: string
          example: Yêu cầu không hợp lệ
        error:
          type: string
          example: Bad Request
...
    Submission:
      type: object
      properties:
        id:
          type: string
          example: s_5f2a
        exerciseId:
          type: string
          example: e3
        userId:
          type: string
          example: u_mina01
        prUrl:
          type: string
          example: github.com/acme/web/pull/468
        status:
          type: string
          enum: [submitted, approved, changes]
          example: submitted
        submittedAt:
          type: string
          format: date-time
          example: 2026-06-15T11:25:00Z
    SubmissionFeedItem:
      type: object
      properties:
        id:
          type: string
          example: s1
        user:
          type: object
          properties:
            id:
              type: string
              example: u_mina01
            name:
              type: string
              example: Mina Okonkwo
            avatarHue:
              type: integer
              example: 162
        exercise:
          type: string
          example: Add a Paginated Users Endpoint
        prUrl:
          type: string
          example: github.com/acme/api/pull/119
        status:
          type: string
          example: submitted
        submittedAt:
          type: string
          format: date-time
          example: 2026-06-15T11:11:00Z
    SubmissionDetail:
      type: object
      properties:
        id:
          type: string
          example: s1
        exerciseId:
          type: string
          example: e2
        exercise:
          type: string
          example: Add a Paginated Users Endpoint
        user:
          type: object
          properties:
            id:
              type: string
              example: u_mina01
            name:
              type: string
              example: Mina Okonkwo
        prUrl:
          type: string
          example: github.com/acme/api/pull/119
        status:
          type: string
          example: submitted
        reviewerId:
          type: string
          nullable: true
          example: null
        reviewNote:
          type: string
          nullable: true
          example: null
        submittedAt:
          type: string
          format: date-time
          example: 2026-06-15T11:11:00Z
        reviewedAt:
          type: string
          format: date-time
          nullable: true
          example: null
    SubmissionHistoryResponse:
      type: object
      properties:
        submissionId:
          type: string
          example: s1
        exerciseId:
          type: string
          example: e2
        history:
          type: array
          items:
            $ref: '#/components/schemas/SubmissionHistoryItem'
    SubmissionHistoryItem:
      type: object
      properties:
        id:
          type: string
          example: sh_102
        prUrl:
          type: string
          example: github.com/acme/api/pull/125
        status:
          type: string
          enum: [submitted, approved, changes]
          example: submitted
        submittedAt:
          type: string
          format: date-time
          example: 2026-06-16T10:15:00Z
        reviewerId:
          type: string
          nullable: true
          example: null
        reviewNote:
          type: string
          nullable: true
          example: null
        reviewedAt:
          type: string
          format: date-time
          nullable: true
          example: null
...